requirements.txt
Django==3.1.7
black==20.8b1
flake8==3.8.4
jupyter==1.0.0
django-extensions==3.1.1
django-ckeditor==6.0.0
requests==2.25.1
djangorestframework==3.12.2
drf-extensions==0.7.0
django-heroku==0.3.1
gunicorn==20.0.4
django-ckeditor==6.0.0
psycopg2
django-filter~=2.4.0
pip~=21.0.1
setuptools~=54.0.0

whitenoise~=5.2.0
